Restructure  - Move data collection to concern and perform moderate cleanup operations in the process  - only resolve exclusions once per execution.
