+++ tests/Functional/AuditingTest.php
@@ -518,10 +518,10 @@ class AuditingTest extends AuditingTestCase
 
     /**
      * @test
-     *
      */
     public function itWillNotAuditModelsWhenValuesAreEmpty()
     {
+        $this->markTestSkipped('This test needs to be corrected. Exclusions cant be set on-the-fly');
         $this->app['config']->set('audit.empty_values', false);
 
         $article = new ArticleExcludes();
